requirements.txt
ConfigArgParse >= 0.13.0
colorama >= 0.3.9
numpy >= 1.16.2
pandas >= 0.24.1
beautifulsoup4 >= 4.7.1
lxml >= 4.3.2
html5lib >= 1.0.1
germalemma >= 0.1.1
nltk >= 3.4
google-cloud-translate >= 1.3.3
tika >= 1.19
regex>=2017.4.5